+++ banking/snapshot.go
@@ -314,3 +314,8 @@ func (e *Engine) restoreAssetActions(ctx context.Context, aa *types.BankingAsset
 	e.bss.changed[assetActionsKey] = true
 	return nil
 }
+
+func (e *Engine) OnEpochRestore(ctx context.Context, ep types.Epoch) {
+	e.log.Debug("epoch restoration notification received", logging.String("epoch", ep.String()))
+	e.currentEpoch = ep.Seq
+}