+++ epochtime/service.go
@@ -24,7 +24,8 @@ type Svc struct {
 	length time.Duration
 	epoch  types.Epoch
 
-	listeners []func(context.Context, types.Epoch)
+	listeners        []func(context.Context, types.Epoch) // for when the epoch state changes
+	restoreListeners []func(context.Context, types.Epoch) // for when the epoch has been restored from a snapshot
 
 	log *logging.Logger
 
@@ -186,10 +187,12 @@ func (s *Svc) fastForward(ctx context.Context) {
 	s.onTick(ctx, tt)
 }
 
-// NotifyOnEpoch allows other services to register a callback function
-// which will be called once we enter a new epoch.
-func (s *Svc) NotifyOnEpoch(f func(context.Context, types.Epoch)) {
+// NotifyOnEpoch allows other services to register 2 callback functions.
+// The first will be called once we enter or leave a new epoch, and the second
+// will be called when the epochserivce has been restored from a snapshot.
+func (s *Svc) NotifyOnEpoch(f func(context.Context, types.Epoch), r func(context.Context, types.Epoch)) {
 	s.listeners = append(s.listeners, f)
+	s.restoreListeners = append(s.restoreListeners, r)
 }
 
 func (s *Svc) notify(ctx context.Context, e types.Epoch) {