+++ execution/mocks/epoch_service_mock.go
@@ -36,13 +36,13 @@ func (m *MockEpochEngine) EXPECT() *MockEpochEngineMockRecorder {
 }
 
 // NotifyOnEpoch mocks base method.
-func (m *MockEpochEngine) NotifyOnEpoch(arg0 func(context.Context, types.Epoch)) {
+func (m *MockEpochEngine) NotifyOnEpoch(arg0, arg1 func(context.Context, types.Epoch)) {
 	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "NotifyOnEpoch", arg0)
+	m.ctrl.Call(m, "NotifyOnEpoch", arg0, arg1)
 }
 
 // NotifyOnEpoch indicates an expected call of NotifyOnEpoch.
-func (mr *MockEpochEngineMockRecorder) NotifyOnEpoch(arg0 interface{}) *gomock.Call {
+func (mr *MockEpochEngineMockRecorder) NotifyOnEpoch(arg0, arg1 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "NotifyOnEpoch", reflect.TypeOf((*MockEpochEngine)(nil).NotifyOnEpoch), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "NotifyOnEpoch", reflect.TypeOf((*MockEpochEngine)(nil).NotifyOnEpoch), arg0, arg1)
 }