+++ liquidity/supplied/mocks/risk_model_mock.go
@@ -7,7 +7,6 @@ package mocks
 import (
 	reflect "reflect"
 
-	num "code.vegaprotocol.io/vega/types/num"
 	gomock "github.com/golang/mock/gomock"
 	decimal "github.com/shopspring/decimal"
 )
@@ -50,7 +49,7 @@ func (mr *MockRiskModelMockRecorder) GetProjectionHorizon() *gomock.Call {
 }
 
 // ProbabilityOfTrading mocks base method.
-func (m *MockRiskModel) ProbabilityOfTrading(arg0, arg1 *num.Uint, arg2, arg3, arg4 decimal.Decimal, arg5, arg6 bool) decimal.Decimal {
+func (m *MockRiskModel) ProbabilityOfTrading(arg0, arg1, arg2, arg3, arg4 decimal.Decimal, arg5, arg6 bool) decimal.Decimal {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "ProbabilityOfTrading", arg0, arg1, arg2, arg3, arg4, arg5, arg6)
 	ret0, _ := ret[0].(decimal.Decimal)